[NOD-781] fix: refactored config cache service
